fix(dimsum): Work with V2 quay response
Quay now responds with V2 docker registry responses by default and while
V1 compatible (ish) responses can still be had, the format is differnt
enough that the package we were using to unmarshal them err's out. This
commit adds a new QuayResponse struct to unmarshal the new V1 responses
into and otherwise replicates the previous dimsum behavior. However, the
image tag is no longer returned to the dimsum client as part of the
response. This should not be a problem though, as the image tag is a
required parameter to query dimsum in the first place, i.e. a
client/user should already have this info.
https://skuidify.atlassian.net/browse/DEVOPS-1279